services/api-gateway/requirements.txt
# api-gateway Python dependencies
fastapi>=0.111.0
uvicorn>=0.30.0
asyncpg>=0.29.0
httpx>=0.27.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
prometheus-fastapi-instrumentator==7.0.0
